requirements.txt
scikit-learn
numpy
matplotlib
pillow
scipy
rawpy
scikit-image
torch
torchvision
pynng
pyyaml